move parallel to comparisons
